# openapi.yaml
...
info:
    title: mahresources
    version: "1.0"
openapi: 3.0.0
paths:
...
    /v1/auth/login:
        post:
            description: Exchanges a username/password for a session cookie. Only meaningful when auth is enabled.
            operationId: login
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                                username:
                                    type: string
                            type: object
                    application/x-www-form-urlencoded:
                        schema:
                            properties:
                                password:
                                    type: string
                                username:
                                    type: string
                            type: object
                required: true
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Authenticate and start a session
            tags:
                - auth
...
    /v1/user:
        get:
            operationId: getUser
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Get a user account (admin)
            tags:
                - users
        post:
            operationId: updateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                disabled:
                                    type: boolean
                                displayName:
                                    type: string
                                id:
                                    type: integer
                                password:
                                    type: string
                                role:
                                    type: string
                                scopeGroupId:
                                    nullable: true
                                    type: integer
                                username:
                                    type: string
                            type: object
                    application/x-www-form-urlencoded:
                        schema:
                            properties:
                                disabled:
                                    type: boolean
                                displayName:
                                    type: string
                                id:
                                    type: integer
                                password:
                                    type: string
                                role:
                                    type: string
                                scopeGroupId:
                                    nullable: true
                                    type: integer
                                username:
                                    type: string
                            type: object
                required: true
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Update a user account (admin)
            tags:
                - users
    /v1/user/delete:
        post:
            operationId: deleteUser
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Delete a user account (admin)
            tags:
                - users
    /v1/users:
        get:
            operationId: listUsers
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: List user accounts (admin)
            tags:
                - users
        post:
            operationId: createUser
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                disabled:
                                    type: boolean
                                displayName:
                                    type: string
                                id:
                                    type: integer
                                password:
                                    type: string
                                role:
                                    type: string
                                scopeGroupId:
                                    nullable: true
                                    type: integer
                                username:
                                    type: string
                            type: object
                    application/x-www-form-urlencoded:
                        schema:
                            properties:
                                disabled:
                                    type: boolean
                                displayName:
                                    type: string
                                id:
                                    type: integer
                                password:
                                    type: string
                                role:
                                    type: string
                                scopeGroupId:
                                    nullable: true
                                    type: integer
                                username:
                                    type: string
                            type: object
                required: true
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Create a user account (admin)
            tags:
                - users
servers:
    - description: API Version 1
      url: /v1
...